switched from floating point engine sizes and powers to integers to match seed data

+++ Eatech.FleetManager.ApplicationCore/Entities/Car.cs
@@ -16,8 +16,8 @@ namespace Eatech.FleetManager.ApplicationCore.Entities
 
         public DateTime? InspectionDate { get; set; }
 
-        public float? EngineSize { get; set; }
+        public int? EngineSize { get; set; }
 
-        public float? EnginePower { get; set; }
+        public int? EnginePower { get; set; }
     }
 }
